pyproject.toml
[tool.poetry]
name = "bridge"
version = "0.1.0"
description = ""
authors = ["Nathan Crock <mathnathan@gmail.com>", "erlebach <gordon.erlebach@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12.*,<3.13"
addict = "2.4.0"
annotated-types = "0.7.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
docker-pycreds = "0.4.0"
exceptiongroup = "1.2.2"
filelock = "3.16.1"
fsspec = "2024.9.0"
gitdb = "4.0.11"
gitpython = "3.1.43"
idna = "3.10"
iniconfig = "2.0.0"
jinja2 = "3.1.4"
joblib = "1.4.2"
markupsafe = "3.0.1"
mpmath = "1.3.0"
networkx = "3.4.1"
nltk = "3.8.1"
numpy = "1.26.0"
packaging = "24.1"
pandas = "2.1.2"
platformdirs = "4.3.6"
pluggy = "1.5.0"
protobuf = "5.28.2"
psutil = "6.0.0"
pydantic = "2.9.2"
pytest = "8.3.3"
pytest-mock = "3.14.0"
python-dateutil = "2.9.0.post0"
python-stdnum = "1.20"
pytz = "2024.2"
pyyaml = "6.0.2"
regex = "2024.9.11"
requests = "2.32.3"
sentry-sdk = "2.16.0"
setproctitle = "1.3.3"
six = "1.16.0"
smmap = "5.0.1"
sympy = "1.13.3"
tomli = "2.0.2"
torch = "2.4.1"
tqdm = "4.66.5"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
wandb = "0.18.3"

triton = {version = "3.0.0", markers = "sys_platform == 'linux'"}

# NVIDIA CUDA and other dependencies
nvidia-cublas-cu12 = {version = "12.1.3.1", markers = "sys_platform == 'linux'"}
nvidia-cuda-cupti-cu12 = {version = "12.1.105", markers = "sys_platform == 'linux'"}
nvidia-cuda-nvrtc-cu12 = {version = "12.1.105", markers = "sys_platform == 'linux'"}
nvidia-cuda-runtime-cu12 = {version = "12.1.105", markers = "sys_platform == 'linux'"}
nvidia-cudnn-cu12 = {version = "9.1.0.70", markers = "sys_platform == 'linux'"}
nvidia-cufft-cu12 = {version = "11.0.2.54", markers = "sys_platform == 'linux'"}
nvidia-curand-cu12 = {version = "10.3.2.106", markers = "sys_platform == 'linux'"}
nvidia-cusolver-cu12 = {version = "11.4.5.107", markers = "sys_platform == 'linux'"}
nvidia-cusparse-cu12 = {version = "12.1.0.106", markers = "sys_platform == 'linux'"}
nvidia-nccl-cu12 = {version = "2.20.5", markers = "sys_platform == 'linux'"}
nvidia-nvjitlink-cu12 = {version = "12.6.77", markers = "sys_platform == 'linux'"}
nvidia-nvtx-cu12 = {version = "12.1.105", markers = "sys_platform == 'linux'"}

# Git dependency
Traindata = {git = "https://github.com/mathnathan/Traindata.git"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
